noc_file_dialog: Specify timestamp to focus dialog for GTK
GNOME has a "focus stealing prevention" feature that will open new
windows in the background to prevent the keyboard focus from being
stolen from a user, and show an "<Application> is ready" notification.

This feature currently impacts file dialog creation resulting in the
dialog being confusingly spawned in the background on every creation
after the first.

This patch adds a call to gtk_window_present_with_time() with the
current time, enforcing that the dialog always makes it to the top of
the window stacking order and receives focus.